+++ src/Stormancer.Plugins/Party/Stormancer.Server.Plugins.Party/PartyService.cs
@@ -69,6 +69,7 @@ namespace Stormancer.Server.Plugins.Party
         private readonly StormancerPartyPlatformSupport _stormancerPartyPlatformSupport;
         private readonly InvitationCodeService invitationCodes;
         private readonly PartyLuceneDocumentStore partyDocumentsStore;
+        private readonly PartyConfigurationService partyConfigurationService;
 
         public IReadOnlyDictionary<SessionId, PartyMember> PartyMembers => _partyState.PartyMembers;
 
@@ -90,7 +91,8 @@ namespace Stormancer.Server.Plugins.Party
             IEnumerable<IPartyPlatformSupport> platformSupports,
             StormancerPartyPlatformSupport stormancerPartyPlatformSupport,
             InvitationCodeService invitationCodes,
-            PartyLuceneDocumentStore partyDocumentsStore
+            PartyLuceneDocumentStore partyDocumentsStore,
+            PartyConfigurationService partyConfigurationService
         )
         {
             _handlers = handlers;
@@ -106,6 +108,7 @@ namespace Stormancer.Server.Plugins.Party
             _stormancerPartyPlatformSupport = stormancerPartyPlatformSupport;
             this.invitationCodes = invitationCodes;
             this.partyDocumentsStore = partyDocumentsStore;
+            this.partyConfigurationService = partyConfigurationService;
             ApplySettings(configuration.Settings);
         }
 
@@ -323,7 +326,16 @@ namespace Stormancer.Server.Plugins.Party
                 if (_partyState.PartyMembers.TryRemove(args.Peer.SessionId, out var partyUser))
                 {
                     Log(LogLevel.Trace, "OnDisconnected", $"Member left the party, reason: {args.Reason}", args.Peer.SessionId, partyUser.UserId);
-                    await TryCancelPendingGameFinder();
+
+                    var handlers = _handlers();
+                    var partyResetCtx = new PartyMemberReadyStateResetContext(PartyMemberReadyStateResetEventType.PartyMembersListUpdated, _scene);
+                    partyConfigurationService.ShouldResetPartyMembersReadyState(partyResetCtx);
+                    await handlers.RunEventHandler(h => h.OnPlayerReadyStateReset(partyResetCtx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while processing an 'OnPlayerReadyStateRest' event.", ex));
+
+                    if (partyResetCtx.ShouldReset)
+                    {
+                        await TryCancelPendingGameFinder();
+                    }
 
                     if (_partyState.Settings.PartyLeaderId == partyUser.UserId && _partyState.PartyMembers.Count != 0)
                     {
@@ -436,7 +448,15 @@ namespace Stormancer.Server.Plugins.Party
                 }
                 _partyState.SettingsVersionNumber = newSettingsVersion;
 
-                await TryCancelPendingGameFinder();
+
+                var partyResetCtx = new PartyMemberReadyStateResetContext(PartyMemberReadyStateResetEventType.PartySettingsUpdated, _scene);
+                partyConfigurationService.ShouldResetPartyMembersReadyState(partyResetCtx);
+                await handlers.RunEventHandler(h => h.OnPlayerReadyStateReset(partyResetCtx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while processing an 'OnPlayerReadyStateRest' event.", ex));
+
+                if (partyResetCtx.ShouldReset)
+                {
+                    await TryCancelPendingGameFinder();
+                }
 
                 Dictionary<PartyMember, PartySettingsUpdateDto> updates = _partyState.PartyMembers.Values.ToDictionary(m => m, _ => new PartySettingsUpdateDto(_partyState));
 
@@ -516,7 +536,14 @@ namespace Stormancer.Server.Plugins.Party
                 }
                 _partyState.SettingsVersionNumber = newSettingsVersion;
 
-                await TryCancelPendingGameFinder();
+                var partyResetCtx = new PartyMemberReadyStateResetContext(PartyMemberReadyStateResetEventType.PartySettingsUpdated, _scene);
+                partyConfigurationService.ShouldResetPartyMembersReadyState(partyResetCtx);
+                await handlers.RunEventHandler(h => h.OnPlayerReadyStateReset(partyResetCtx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while processing an 'OnPlayerReadyStateRest' event.", ex));
+
+                if (partyResetCtx.ShouldReset)
+                {
+                    await TryCancelPendingGameFinder();
+                }
 
                 Dictionary<PartyMember, PartySettingsUpdateDto> updates = _partyState.PartyMembers.Values.ToDictionary(m => m, _ => new PartySettingsUpdateDto(_partyState));
 
@@ -570,6 +597,15 @@ namespace Stormancer.Server.Plugins.Party
                     throw new ClientException(SettingsOutdatedError);
                 }
 
+                var updatingCtx = new UpdatingPlayerReadyStateContext(this, user, _scene);
+
+                await handlers.RunEventHandler(h => h.OnUpdatingPlayerReadyState(updatingCtx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running 'OnUpdatingPlayerReadyState'", ex));
+
+                if (!updatingCtx.Accept)
+                {
+                    throw new ClientException(string.IsNullOrEmpty(updatingCtx.ErrorId) ? "party.ready.updateDenied" : updatingCtx.ErrorId);
+                }
+
                 user.StatusInParty = partyUserStatus.DesiredStatus;
                 Log(LogLevel.Trace, "UpdateGameFinderPlayerStatus", $"Updated user status, new value: {partyUserStatus}", user.Peer.SessionId, user.UserId);
 
@@ -577,7 +613,7 @@ namespace Stormancer.Server.Plugins.Party
                 update.UserStatus.Add(new PartyMemberStatusUpdate { UserId = userId, Status = user.StatusInParty });
                 await BroadcastStateUpdateRpc(BatchStatusUpdate.Route, update);
 
-                var eventHandlerCtx = new PlayerReadyStateContext(this, user);
+                var eventHandlerCtx = new PlayerReadyStateContext(this, user, _scene);
                 await handlers.RunEventHandler(h => h.OnPlayerReadyStateChanged(eventHandlerCtx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running OnPlayerReadyStateChanged", ex));
 
                 bool shouldLaunchGameFinderRequest = false;
@@ -632,8 +668,29 @@ namespace Stormancer.Server.Plugins.Party
                     ThrowNoSuchMemberError(userId);
                 }
 
+                await using var scope = _scene.CreateRequestScope();
+                var handlers = scope.Resolve<IEnumerable<IPartyEventHandler>>();
+
+                var ctx = new UpdatingPartyMemberDataContext(partyUser, data, _scene);
+                await handlers.RunEventHandler(h => h.OnUpdatingPartyMemberData(ctx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while running the event 'OnUpdatingPartyMemberData'.", ex));
+
+                if (!ctx.IsUpdateValid)
+                {
+                    throw new ClientException(ctx.Error ?? "party.invalidMemberData");
+                }
                 partyUser.UserData = data;
                 Log(LogLevel.Trace, "UpdatePartyUserData", "Updated user data", new { partyUser.Peer.SessionId, partyUser.UserId, UserData = data });
+
+                var partyResetctx = new PartyMemberReadyStateResetContext(PartyMemberReadyStateResetEventType.PartyMemberDataUpdated, _scene);
+                partyConfigurationService.ShouldResetPartyMembersReadyState(partyResetctx);
+                await handlers.RunEventHandler(h => h.OnPlayerReadyStateReset(partyResetctx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while processing an 'OnPlayerReadyStateRest' event.", ex));
+
+                if (partyResetctx.ShouldReset)
+                {
+                    await TryCancelPendingGameFinder();
+                }
+
+
 
                 await BroadcastStateUpdateRpc(PartyMemberDataUpdate.Route, new PartyMemberDataUpdate { UserId = userId, UserData = partyUser.UserData });
             });
@@ -681,7 +738,16 @@ namespace Stormancer.Server.Plugins.Party
                     {
                         throw new ClientException(CannotKickLeaderError);
                     }
-                    await TryCancelPendingGameFinder();
+                    var handlers = _handlers();
+                    var ctx = new PartyMemberReadyStateResetContext(PartyMemberReadyStateResetEventType.PartyMembersListUpdated, _scene);
+                    partyConfigurationService.ShouldResetPartyMembersReadyState(ctx);
+                    await handlers.RunEventHandler(h => h.OnPlayerReadyStateReset(ctx), ex => _logger.Log(LogLevel.Error, "party", "An error occured while processing an 'OnPlayerReadyStateRest' event.", ex));
+
+                    if (ctx.ShouldReset)
+                    {
+                        await TryCancelPendingGameFinder();
+                    }
+
 
                     _partyState.PartyMembers.TryRemove(partyUser.Peer.SessionId, out _);